requirements.txt
# Core automation
playwright>=1.40.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
oauth2client>=4.1.3

# Notifications
python-telegram-bot>=20.0
requests>=2.31.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0